feat(acp): 支持通过 WebSocket 消息延迟初始化沙箱并增加配置批量注入功能
- 新增 deferredInitMap 用于暂存未携带 session/config 消息的 WebSocket 连接参数
- 支持等待前端发送 session/config 消息后再启动沙箱初始化流水线
- ConfigInjectionPhase 改用 tar.gz 包批量注入配置文件，避免 JSON 序列化导致的 Unicode 规范化问题
- 实现 SandboxProvider.extractArchive 接口，通过 Sidecar HTTP API 上传 tar.gz 压缩包并解压
- K8sSandboxProvider 和 LocalSandboxProvider 实现 extractArchive 方法
- 增加对配置文件注入后抽样读回验证，确保文件正确解压且不为空
- PodReuseManager 新增等待 SLB 后端健康检查通过逻辑，确保流量可达 Pod 的 sidecar
- AcpWebSocketHandler 优化消息处理逻辑，支持兼容旧客户端同时对新客户端延迟启动pipeline
- 修复 SkillPackageServiceImpl 中编码内容长度计算冗余问题
- README 文档补充 Agent Skill 相关描述及 ServiceAccount 绑定 Role 的声明
- 测试覆盖改进，新增 ConfigInjectionPhase 相关单元测试，覆盖 tar.gz 解压及异常场景处理

+++ himarket-server/src/main/java/com/alibaba/himarket/service/impl/SkillPackageServiceImpl.java
@@ -222,8 +222,7 @@ public class SkillPackageServiceImpl implements SkillPackageService {
                 .ifPresentOrElse(
                         skillFile -> {
                             skillFile.setContent(content);
-                            skillFile.setSize(
-                                    (long) content.getBytes(StandardCharsets.UTF_8).length);
+                            skillFile.setSize(content.getBytes(StandardCharsets.UTF_8).length);
                             skillFileRepository.save(skillFile);
                         },
                         () -> {
@@ -233,10 +232,7 @@ public class SkillPackageServiceImpl implements SkillPackageService {
                                             .path("SKILL.md")
                                             .encoding("utf-8")
                                             .content(content)
-                                            .size(
-                                                    (long)
-                                                            content.getBytes(StandardCharsets.UTF_8)
-                                                                    .length)
+                                            .size(content.getBytes(StandardCharsets.UTF_8).length)
                                             .build();
                             skillFileRepository.save(skillFile);
                         });
